+++ src/main/java/com/github/liblevenshtein/distance/MemoizedMergeAndSplit.java
@@ -1,5 +1,7 @@
 package com.github.liblevenshtein.distance;
 
+import lombok.val;
+
 import com.github.liblevenshtein.collection.SymmetricImmutablePair;
 
 /**
@@ -18,7 +20,7 @@ public class MemoizedMergeAndSplit extends AbstractMemoized {
   @Override
   @SuppressWarnings("checkstyle:finalparameters")
   public int memoizedDistance(String v, String w) {
-    final SymmetricImmutablePair key = new SymmetricImmutablePair<String>(v, w);
+    val key = new SymmetricImmutablePair<String>(v, w);
 
     int distance = memo.getInt(key);
     if (distance != DEFAULT_RETURN_VALUE) {